Split budget table model logic
